replace int with more suitable types

+++ src/common.h
@@ -7,14 +7,16 @@
 #ifndef M_TRIE_COMMON_H
 #define M_TRIE_COMMON_H
 
+#include <stdint.h>
+
 #include "m_trie.h"
 #include "node.h"
 
 
-int check(const struct m_trie* tr, const uint8_t* key, const uint32_t len);
-int locate(const struct m_trie* tr,
-           const uint8_t* key,
-           const uint32_t len,
-           struct node** nd);
+uint8_t check(const struct m_trie* tr, const uint8_t* key, const uint32_t len);
+uint8_t locate(const struct m_trie* tr,
+               const uint8_t* key,
+               const uint32_t len,
+               struct node** nd);
 
 #endif
